requirements.txt
# Для пользователей с обычным Python (не Kali):
# pip install -r requirements.txt
requests>=2.28.0
python-whois>=0.8.0
